+++ gazon-service/src/test/java/com/gitlab/controller/ReviewRestControllerIT.java
@@ -14,13 +14,18 @@ import java.time.LocalDate;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.testcontainers.shaded.org.hamcrest.CoreMatchers.equalTo;
+import static org.testcontainers.shaded.org.hamcrest.MatcherAssert.assertThat;
 
-public class ReviewRestControllerIT extends AbstractIntegrationTest {
+class ReviewRestControllerIT extends AbstractIntegrationTest {
 
     private static final String REVIEW_URN = "/api/review";
     private static final String REVIEW_URI = URL + REVIEW_URN;
@@ -85,6 +90,8 @@ public class ReviewRestControllerIT extends AbstractIntegrationTest {
     @Test
     void should_update_review_by_id() throws Exception {
         long id = 1L;
+        int numberOfEntitiesExpected = reviewService.findAll().size();
+
         ReviewDto reviewDto = generateReviewDto();
 
         String jsonReviewDto = objectMapper.writeValueAsString(reviewDto);
@@ -98,7 +105,9 @@ public class ReviewRestControllerIT extends AbstractIntegrationTest {
                         .accept(MediaType.APPLICATION_JSON))
                 .andDo(print())
                 .andExpect(status().isOk())
-                .andExpect(content().json(expected));
+                .andExpect(content().json(expected))
+                .andExpect(result -> assertThat(reviewService.findAll().size(),
+                        equalTo(numberOfEntitiesExpected)));
     }
 
     @Test
@@ -116,7 +125,6 @@ public class ReviewRestControllerIT extends AbstractIntegrationTest {
                 .andExpect(status().isNotFound());
     }
 
-
     @Test
     void should_delete_review_by_id() throws Exception {
         long id = 3L;
@@ -127,7 +135,6 @@ public class ReviewRestControllerIT extends AbstractIntegrationTest {
                 .andDo(print())
                 .andExpect(status().isNotFound());
     }
-
 
     @Test
     void should_get_images_ids_by_review_id() throws Exception {
@@ -146,7 +153,6 @@ public class ReviewRestControllerIT extends AbstractIntegrationTest {
                 .andExpect(content().json(expected));
     }
 
-
     @Test
     void should_create_multiple_reviewImages_by_review_id() throws Exception {
         long id = 1L;
@@ -160,7 +166,6 @@ public class ReviewRestControllerIT extends AbstractIntegrationTest {
                         .accept(MediaType.ALL))
                 .andExpect(status().isCreated());
     }
-
 
     @Test
     void should_delete_all_reviewImages_by_review_id() throws Exception {